move alterTableStats logic to HiveExternalCatalog
