+++ sql/core/src/main/scala/org/apache/spark/sql/execution/command/AnalyzeTableCommand.scala
@@ -31,9 +31,6 @@ import org.apache.spark.sql.execution.datasources.LogicalRelation
 /**
  * Analyzes the given table in the current database to generate statistics, which will be
  * used in query optimizations.
- *
- * Right now, it only supports Hive tables and it only updates the size of a Hive table
- * in the Hive metastore.
  */
 case class AnalyzeTableCommand(tableName: String, noscan: Boolean = true) extends RunnableCommand {
